requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
cattrs==25.1.1
celery==5.5.3
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7.1
cryptography==45.0.5
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.4
django-admin-interface==0.30.1
django-cleanup==9.0.0
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
django-schema-graph==3.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
Faker==37.5.3
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
locust-cloud==1.26.3
MarkupSafe==3.0.2
msgpack==1.1.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.13.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.27.1
ruff==0.12.8
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.3
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.10.0
wsproto==1.2.0
zibal_payment==1.0.2
zope.event==5.1.1
zope.interface==7.2
